// src/app/models/studentModel.ts
export class studentModel{
    id?: number | undefined;
    person_id?: number | undefined;
    career_id?: number | undefined;
    institutional_email?: string | undefined;
    pay_method?: string | undefined;
    admission_date?: string | undefined;
    guardian_name?: string | undefined;
    home_phone?: string | undefined;
    status?: string | undefined;
    person_name?: string | undefined;
    career_name?: string | undefined;

 //------
    phase_id?: number | undefined;
    link?: string | undefined;

  }